requirements.txt
tensorflow-addons==0.13.0
marshmallow==3.12.1
marshmallow-enum==1.5.1
typing-inspect==0.7.1
pytest==6.2.4
grpcio==1.31.0
grpcio-tools==1.31.0
python-dotenv==0.18.0
